requirements.txt
aiogram
langchain
langchain-core
langchain-openai
langchain-community
openai
requests
python-dotenv